refactor: - dedupe flash / highlighting logic - codemirror logic now lives only in codemirror package - remove old highlighting logic - use codemirror package in react package - cleanup CodeMirror6.jsx - pull setMiniLocations into useHighlighting - migrate MiniRepl, nano-repl + Repl to new highlighting
