+++ packages/codemirror/codemirror.mjs
@@ -1,11 +1,12 @@
-import { EditorState } from '@codemirror/state';
-import { EditorView, keymap, Decoration, lineNumbers, highlightActiveLineGutter } from '@codemirror/view';
 import { defaultKeymap } from '@codemirror/commands';
-import { syntaxHighlighting, defaultHighlightStyle } from '@codemirror/language';
 import { javascript } from '@codemirror/lang-javascript';
-import { StateField, StateEffect } from '@codemirror/state';
+import { defaultHighlightStyle, syntaxHighlighting } from '@codemirror/language';
+import { EditorState } from '@codemirror/state';
+import { EditorView, highlightActiveLineGutter, keymap, lineNumbers } from '@codemirror/view';
+import { Drawer, repl } from '@strudel.cycles/core';
+import { flashField, flash } from './flash.mjs';
+import { highlightExtension, highlightMiniLocations } from './highlight.mjs';
 import { oneDark } from './themes/one-dark';
-import { repl, Drawer } from '@strudel.cycles/core';
 
 // https://codemirror.net/docs/guide/
 export function initEditor({ initialCode = '', onChange, onEvaluate, onStop, theme = oneDark, root }) {
@@ -15,7 +16,7 @@ export function initEditor({ initialCode = '', onChange, onEvaluate, onStop, the
       theme,
       javascript(),
       lineNumbers(),
-      highlightField,
+      highlightExtension,
       highlightActiveLineGutter(),
       syntaxHighlighting(defaultHighlightStyle),
       keymap.of(defaultKeymap),
@@ -40,93 +41,6 @@ export function initEditor({ initialCode = '', onChange, onEvaluate, onStop, the
   });
 }
 
-//
-//  highlighting
-//
-
-export const setHighlights = StateEffect.define();
-export const highlightField = StateField.define({
-  create() {
-    return Decoration.none;
-  },
-  update(highlights, tr) {
-    try {
-      for (let e of tr.effects) {
-        if (e.is(setHighlights)) {
-          const { haps } = e.value;
-          const marks =
-            haps
-              .map((hap) =>
-                (hap.context.locations || []).map(({ start, end }) => {
-                  // const color = hap.context.color || e.value.color || '#FFCA28';
-                  let from = tr.newDoc.line(start.line).from + start.column;
-                  let to = tr.newDoc.line(end.line).from + end.column;
-                  const l = tr.newDoc.length;
-                  if (from > l || to > l) {
-                    return; // dont mark outside of range, as it will throw an error
-                  }
-                  const mark = Decoration.mark({
-                    attributes: { style: `outline: 2px solid #FFCA28;` },
-                  });
-                  return mark.range(from, to);
-                }),
-              )
-              .flat()
-              .filter(Boolean) || [];
-          highlights = Decoration.set(marks, true);
-        }
-      }
-      return highlights;
-    } catch (err) {
-      // console.warn('highlighting error', err);
-      return Decoration.set([]);
-    }
-  },
-  provide: (f) => EditorView.decorations.from(f),
-});
-
-// helper to simply trigger highlighting for given haps
-export function highlightHaps(view, haps) {
-  view.dispatch({ effects: setHighlights.of({ haps }) });
-}
-
-//
-// flash
-//
-
-export const setFlash = StateEffect.define();
-const flashField = StateField.define({
-  create() {
-    return Decoration.none;
-  },
-  update(flash, tr) {
-    try {
-      for (let e of tr.effects) {
-        if (e.is(setFlash)) {
-          if (e.value && tr.newDoc.length > 0) {
-            const mark = Decoration.mark({ attributes: { style: `background-color: #FFCA2880` } });
-            flash = Decoration.set([mark.range(0, tr.newDoc.length)]);
-          } else {
-            flash = Decoration.set([]);
-          }
-        }
-      }
-      return flash;
-    } catch (err) {
-      console.warn('flash error', err);
-      return flash;
-    }
-  },
-  provide: (f) => EditorView.decorations.from(f),
-});
-
-export const flash = (view, ms = 200) => {
-  view.dispatch({ effects: setFlash.of(true) });
-  setTimeout(() => {
-    view.dispatch({ effects: setFlash.of(false) });
-  }, ms);
-};
-
 export class StrudelMirror {
   constructor(options) {
     const { root, initialCode = '', onDraw, drawTime = [-2, 2], prebake, ...replOptions } = options;
@@ -134,7 +48,7 @@ export class StrudelMirror {
 
     this.drawer = new Drawer((haps, time) => {
       const currentFrame = haps.filter((hap) => time >= hap.whole.begin && time <= hap.endClipped);
-      this.highlight(currentFrame);
+      this.highlight(currentFrame, time);
       onDraw?.(haps, time, currentFrame);
     }, drawTime);
 
@@ -193,7 +107,7 @@ export class StrudelMirror {
   flash(ms) {
     flash(this.editor, ms);
   }
-  highlight(haps) {
-    highlightHaps(this.editor, haps);
+  highlight(haps, time) {
+    highlightMiniLocations(this.editor.view, time, haps);
   }
 }
