+++ packages/codemirror/index.mjs
@@ -1,0 +1,3 @@
+export * from './codemirror.mjs';
+export * from './highlight.mjs';
+export * from './flash.mjs';
